requirements.txt
fastapi
uvicorn[standard]
motor
sendgrid
python-jose[cryptography]
python-multipart
passlib
jinja2
pillow